Refactor into skeltal package
