+++ src/main/java/skeltal/Deadline.java
@@ -1,18 +1,19 @@
+package skeltal;;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-public class Event extends Task {
+
+public class Deadline extends Task {
     private String time;
 
-    public Event(String rawTime) throws SkeltalException {
+    public Deadline(String rawTime) throws SkeltalException{
         super(rawTime.split("/", 2)[0]);
         String[] procTime = rawTime.split("/", 2);
-        String time;
-
         if (procTime.length == 1) {
-            throw new SkeltalException("OOPS! The description of an event cannot be empty!");
+            throw new SkeltalException("OOPS! The description of a deadline cannot be empty!");
         }
 
+        String time;
         try {
             LocalDate date = LocalDate.parse(procTime[1], DateTimeFormatter.ofPattern("dd-MM-yyyy"));
             time = date.format(DateTimeFormatter.ofPattern("dd MMM yy"));
@@ -23,17 +24,16 @@ public class Event extends Task {
     }
 
     private String formatTime() {
-        String formatTime = "(" + this.time + ")";
-        return formatTime;
+        return "(" + this.time + ")";
     }
 
     @Override
     public String store() {
-        return "E | " + super.store() + "| " + time;
+        return "D | " + super.store() + "| " + this.time;
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + formatTime();
+        return "[D]" + super.toString() + formatTime();
     }
 }
