+++ src/main/java/skeltal/Event.java
@@ -1,18 +1,19 @@
+package skeltal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-
-public class Deadline extends Task {
+public class Event extends Task {
     private String time;
 
-    public Deadline(String rawTime) throws SkeltalException{
+    public Event(String rawTime) throws SkeltalException {
         super(rawTime.split("/", 2)[0]);
         String[] procTime = rawTime.split("/", 2);
+        String time;
+
         if (procTime.length == 1) {
-            throw new SkeltalException("OOPS! The description of a deadline cannot be empty!");
+            throw new SkeltalException("OOPS! The description of an event cannot be empty!");
         }
 
-        String time;
         try {
             LocalDate date = LocalDate.parse(procTime[1], DateTimeFormatter.ofPattern("dd-MM-yyyy"));
             time = date.format(DateTimeFormatter.ofPattern("dd MMM yy"));
@@ -23,17 +24,16 @@ public class Deadline extends Task {
     }
 
     private String formatTime() {
-        String formatTime = "(" + this.time + ")";
-        return formatTime;
+        return "(" + this.time + ")";
     }
 
     @Override
     public String store() {
-        return "D | " + super.store() + "| " + this.time;
+        return "E | " + super.store() + "| " + time;
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + formatTime();
+        return "[E]" + super.toString() + formatTime();
     }
 }
