+++ src/main/java/skeltal/Skeltal.java
@@ -1,3 +1,4 @@
+package skeltal;
 import java.util.Scanner;
 
 public class Skeltal {
