+++ src/main/java/skeltal/SkeltalException.java
@@ -1,3 +1,4 @@
+package skeltal;
 public class SkeltalException extends Exception {
     public SkeltalException(String errorMessage) {
         super(errorMessage);
