+++ src/main/java/skeltal/Storage.java
@@ -1,3 +1,4 @@
+package skeltal;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Scanner;
