+++ src/main/java/skeltal/Task.java
@@ -1,3 +1,4 @@
+package skeltal;
 public class Task {
 
     private boolean completed;
